LayoutTests/media/modern-media-controls/placard-support/placard-support-pip.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<style type="text/css" media="screen">
    
    video, #host {
        position: absolute;
        top: 0;
        left: 0;
    }

    video {
        width: 800px;
        height: 240px;
    }
    
</style>
<video src="../../content/test.mp4" autoplay></video>
<div id="host"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>PlacardSupport</code> behavior when entering picture-in-picture.");

iconService.directoryPath = "../../../../Source/WebCore/Modules/modern-media-controls/images";

if (window.internals)
    window.internals.settings.setAllowsPictureInPictureMediaPlayback(true);

const container = document.querySelector("div#host");
const media = document.querySelector("video");
const mediaController = createControls(container, media, null);

const button = document.body.appendChild(document.createElement("button"));

button.addEventListener("click", () => {
    media.webkitSetPresentationMode(media.webkitPresentationMode == "inline" ? "picture-in-picture" : "inline");
});

media.addEventListener("webkitpresentationmodechanged", () => {
    if (media.webkitPresentationMode == "picture-in-picture") {
        shouldBeTrue("mediaController.controls.showsPlacard");
        shouldBeTrue("mediaController.controls.children.includes(mediaController.controls.pipPlacard)");
        clickButton();
    } else {
        shouldBeFalse("mediaController.controls.showsPlacard");
        shouldBeFalse("mediaController.controls.children.includes(mediaController.controls.pipPlacard)");
        container.remove();
        media.remove();
        finishJSTest();
    }
});

media.addEventListener("play", clickButton);

function clickButton() {
    const bounds = button.getBoundingClientRect();
    const centerX = bounds.left + bounds.width / 2;
    const centerY = bounds.top + bounds.height / 2;

    eventSender.mouseMoveTo(centerX, centerY);
    eventSender.mouseDown();
    eventSender.mouseUp();
}

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
